+++ Demo/KHDataBindingDemo/KHDataBindingDemo/UserInfoCell.m
@@ -36,7 +36,7 @@
     [self.pairInfo loadImageURL:model.picture.medium imageView:self.imgUserPic placeHolder:nil brokenImage:nil animation:YES];
     
     NSIndexPath *index = [self.pairInfo indexPath];
-    self.lbNumber.text = [NSString stringWithFormat:@"%ld", index.row ];
+    self.lbNumber.text = [NSString stringWithFormat:@"%ld", (long)index.row ];
 }
